+++ archeryutils/handicaps/handicap_functions.py
@@ -26,7 +26,7 @@ import numpy.typing as npt
 
 from archeryutils import rounds, targets
 
-from .handicap_scheme import FloatArray, HandicapScheme
+from .handicap_scheme import HandicapScheme
 from .handicap_scheme_aa import HandicapAA, HandicapAA2
 from .handicap_scheme_agb import HandicapAGB, HandicapAGBold
 
@@ -98,18 +98,18 @@ def handicap_scheme(  # noqa: D417 - Missing argument in docstring (**kwargs)
 
 
 def arrow_score(
-    handicap: FloatArray,
+    handicap: npt.ArrayLike,
     target: targets.Target,
     handicap_sys: str | HandicapScheme,
     arw_d: float | None = None,
-) -> FloatArray:
+) -> npt.NDArray[np.float64]:
     """
     Calculate the average arrow score for a given target and handicap rating.
 
     Parameters
     ----------
-    handicap : FloatArray
-        handicap value to calculate score for
+    handicap : ArrayLike
+        handicap(s) to calculate score for
     target : targets.Target
         A Target class specifying the target to be used
     handicap_sys : str | HandicapScheme
@@ -119,7 +119,7 @@ def arrow_score(
 
     Returns
     -------
-    FloatArray
+    NDArray[np.float64]
         average expected arrow score for this handicap and target
 
     Warnings
@@ -150,7 +150,7 @@ def arrow_score(
 
 
 def score_for_passes(
-    handicap: FloatArray,
+    handicap: npt.ArrayLike,
     rnd: rounds.Round,
     handicap_sys: str | HandicapScheme,
     arw_d: float | None = None,
@@ -161,8 +161,8 @@ def score_for_passes(
 
     Parameters
     ----------
-    handicap : FloatArray
-        handicap value to calculate score for
+    handicap : ArrayLike
+        handicap value(s) to calculate score for
     rnd : rounds.Round
         A Round class specifying the round being shot
     handicap_sys : str | HandicapScheme
@@ -214,7 +214,7 @@ def score_for_passes(
 
 
 def score_for_round(
-    handicap: FloatArray,
+    handicap: npt.ArrayLike,
     rnd: rounds.Round,
     handicap_sys: str | HandicapScheme,
     arw_d: float | None = None,
@@ -225,8 +225,8 @@ def score_for_round(
 
     Parameters
     ----------
-    handicap : FloatArray
-        handicap value to calculate score for
+    handicap : ArrayLike
+        handicap(s) to calculate score for
     rnd : rounds.Round
         A Round class specifying the round being shot
     handicap_sys : str | HandicapScheme
@@ -238,7 +238,7 @@ def score_for_round(
 
     Returns
     -------
-    round_score : FloatArray
+    round_score : NDArray[np.float64]
         average score of the round for this set of parameters
 
     Examples